Reporte novedades añadido con funciones básicas, junto con algunas correcciones en reporte de actividades mineras

+++ database/migrations/2015_06_23_202328_create_activity_reports_table.php
@@ -18,8 +18,8 @@ class CreateActivityReportsTable extends Migration {
 			$table->integer('sub_cost_center_id')->unsigned(); // foreign key
 			$table->integer('employee_id')->unsigned(); // foreign key
 			$table->integer('mining_activity_id')->unsigned(); // foreign key
-			$table->integer('quantity')->default(0);
-			$table->integer('price')->default('0');
+			$table->decimal('quantity', 3, 2)->default(0);
+			$table->bigInteger('price')->default('0');
 			$table->string('comment')->nullable();
 			$table->integer('reported_by')->unsigned(); // foreign key
 			$table->dateTime('reported_at');
